♻️ Export all exceptions in lamindb_setup.errors as in lamindb
