+++ lamindb_setup/core/exceptions.py
@@ -1,10 +1,1 @@
-from __future__ import annotations
-
-
-class DefaultMessageException(Exception):
-    default_message: str | None = None
-
-    def __init__(self, message: str | None = None):
-        if message is None:
-            message = self.default_message
-        super().__init__(message)
+from lamindb_setup.errors import DefaultMessageException  # backwards compatibility
