+++ lamindb_setup/errors.py
@@ -1,0 +1,28 @@
+from __future__ import annotations
+
+
+class DefaultMessageException(Exception):
+    default_message: str | None = None
+
+    def __init__(self, message: str | None = None):
+        if message is None:
+            message = self.default_message
+        super().__init__(message)
+
+
+class ModuleWasntConfigured(Exception):
+    pass
+
+
+# raise if a cloud SQLite instance is already locked
+# ignored by unlock_cloud_sqlite_upon_exception
+class InstanceLockedException(Exception):
+    pass
+
+
+class SettingsEnvFileOutdated(Exception):
+    pass
+
+
+class CannotSwitchDefaultInstance(Exception):
+    pass
